+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.3] - 2020-11-27
+### Changed
+- [Feature] Uploaded games use the name of the build instead of "Untitled" as the title of the game
+- [UI] Minor copy adjustment
+
+### Fixed
+- [Documentation] Fixed incorrect documentation for PublisherUtils.GetFilteredGameTitle()
+
 ## [4.0.2] - 2020-11-25
 ### Changed
 - [UI] Clarified the copy